pyproject.toml
[tool.poetry]
name = "ds-project"
version = "0.1.0"
description = "Inital project structure for UCARE.AI data science team"
authors = ["Vdovenko Anton"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
beautifulsoup4 = "^4.12.2"
ipykernel = "^6.26.0"
requests = "^2.31.0"
pandas = "^2.1.3"
loguru = "^0.7.2"
bs4 = "^0.0.1"
numpy = "^1.26.1"
interrogate = "^1.5.0"
black = "^23.11.0"
telethon = "^1.33.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pre-commit = "^3.3.2"
isort = "^5.12.0"
interrogate = "^1.5.0"
hydra-core = "^1.3.2"
pytest = "^7.3.1"
ruff = "^0.0.271"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''

[tool.isort]
line_length = 88
atomic=true
force_single_line=true
lines_after_imports=2
lines_between_types=1
use_parentheses=true
known_first_party="interrogate"
known_third_party=["attr", "click", "py", "pytest", "setuptools", "tabulate"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-module = false
ignore-property-decorators = false
fail-under = 95
exclude = ["tests/functional/sample", "setup.py", "docs"]
verbose = 0
omit-covered-files = false
quiet = false
whitelist-regex = []
ignore-regex = []
color = true

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
# target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
